user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["ilyasokur <47240755+ilyasokur@users.noreply.github.com>"]
readme = ""

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
psycopg2 = "^2.9.10"
alembic = "^1.14.1"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
httpx = "^0.28.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
requests = "^2.32.3"
bcrypt = "^4.3.0"
cryptography = "^44.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
